pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turborag"
version = "0.1.0"
description = "Compressed vector and graph-augmented retrieval primitives inspired by the TurboRAG specification."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Ratnam Shah" }
]
dependencies = [
  "click>=8.0",
  "numpy>=1.24",
  "scipy>=1.10",
]

[project.optional-dependencies]
embed = [
  "sentence-transformers>=2.5",
]
graph = [
  "networkx>=3.0",
  "leidenalg>=0.10",
  "python-igraph>=0.10",
]
ingest = [
  "pdfminer.six>=20221105",
  "tiktoken>=0.6",
]
serve = [
  "starlette>=0.37",
  "uvicorn>=0.27",
]
mcp = [
  "mcp>=1.0",
]
all = [
  "sentence-transformers>=2.5",
  "networkx>=3.0",
  "leidenalg>=0.10",
  "python-igraph>=0.10",
  "pdfminer.six>=20221105",
  "tiktoken>=0.6",
  "starlette>=0.37",
  "uvicorn>=0.27",
  "mcp>=1.0",
]

dev = [
  "pytest>=8.0",
]

[project.scripts]
turborag = "turborag.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
